Seminar/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyMuPDF==1.23.5
Pillow==10.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0

# pip install -r requirements.txt